data_generator/requirements.txt
kafka-python-ng>=2.2.3   # kafka-python fork with Python 3.12+ support
